+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 [Unreleased](https://github.com/jshwi/pyaud/compare/v3.6.0...HEAD)
 ------------------------------------------------------------------------
 ### Added
+- Logs commencement of audit
 - Adds logger for coverage.xml path
 
 [3.6.0](https://github.com/jshwi/pyaud/releases/tag/v3.6.0) - 2022-01-04